+++ client/package.json
@@ -10,7 +10,7 @@
     "react-dom": "^16.8.6",
     "react-markdown": "^4.0.4",
     "react-moment": "^0.7.0",
-    "react-redux": "^7.0.3",
+    "react-redux": "^7.1.0",
     "react-router-dom": "^5.0.0",
     "react-scripts": "^3.0.1",
     "redux": "^4.0.1",
